+++ nuxeo-drive-client/nxdrive/client/local_client.py
@@ -13,6 +13,7 @@ import unicodedata
 import warnings
 
 from send2trash import send2trash
+from typing import List, Optional, Text, Tuple, Union
 
 from nxdrive.client.base_automation_client import DOWNLOAD_TMP_FILE_PREFIX, \
     DOWNLOAD_TMP_FILE_SUFFIX
@@ -21,11 +22,10 @@ from nxdrive.client.common import BaseClient, DuplicationDisabledError, \
     safe_filename
 from nxdrive.logging_config import get_logger
 from nxdrive.options import Options
-from nxdrive.osi import AbstractOSIntegration
 from nxdrive.utils import guess_digest_algorithm, normalized_path, \
     safe_long_path
 
-if AbstractOSIntegration.is_windows():
+if sys.platform == 'win32':
     import ctypes
     import win32api
     import win32con
@@ -43,7 +43,7 @@ DEDUPED_BASENAME_PATTERN = ur'^(.*)__(\d{1,3})$'
 # Data transfer objects
 
 class FileInfo(object):
-    """Data Transfer Object for file info on the Local FS"""
+    """ Data Transfer Object for file info on the Local FS. """
 
     def __init__(self, root, path, folderish, last_modification_time, **kwargs):
         # Function to check during long-running processing like digest
@@ -60,7 +60,7 @@ class FileInfo(object):
         # NXDRIVE-188: normalize name on the file system if not normalized
         if (os.path.exists(filepath)
                 and normalized_filepath != filepath
-                and not AbstractOSIntegration.is_mac()):
+                and not sys.platform == 'darwin'):
             log.debug('Forcing normalization of %r to %r',
                       filepath, normalized_filepath)
             os.rename(filepath, normalized_filepath)
@@ -87,10 +87,12 @@ class FileInfo(object):
         return u'FileInfo[%s, remote_ref=%s]' % (self.filepath, self.remote_ref)
 
     def get_digest(self, digest_func=None):
+        # type: (Optional[callable]) -> Union[Text, None]
         """ Lazy computation of the digest. """
 
         if self.folderish:
             return None
+
         digest_func = (digest_func
                        if digest_func is not None
                        else self._digest_func)
@@ -116,7 +118,7 @@ class FileInfo(object):
 
 
 class LocalClient(BaseClient):
-    """Client API implementation for the local file system"""
+    """ Client API implementation for the local file system. """
 
     CASE_RENAME_PREFIX = 'driveCaseRename_'
 
@@ -136,17 +138,20 @@ class LocalClient(BaseClient):
     def __repr__(self):
         return ('<{name}'
                 ' base_folder={cls.base_folder!r},'
-                ' duplication_enabled={cls._disable_duplication!r},'
+                ' duplication_disabled={cls._disable_duplication!r},'
                 ' is_case_sensitive={cls._case_sensitive!r}'
                 '>'
                 ).format(name=type(self).__name__, cls=self)
 
     def duplication_enabled(self):
+        # type: () -> bool
         """ Check if de-duplication is enable or not. """
 
         return not self._disable_duplication
 
     def is_case_sensitive(self):
+        # type: () -> bool
+
         if self._case_sensitive is None:
             lock = self.unlock_path(self.base_folder, unlock_parent=False)
             path = tempfile.mkdtemp(prefix='.caseTest_',
@@ -158,19 +163,27 @@ class LocalClient(BaseClient):
 
     @staticmethod
     def is_temp_file(filename):
+        # type: () -> bool
+
         return (filename.startswith(DOWNLOAD_TMP_FILE_PREFIX) and
                 filename.endswith(DOWNLOAD_TMP_FILE_SUFFIX))
 
     def set_readonly(self, ref):
+        # type: (Text) -> None
+
         path = self.abspath(ref)
         self.set_path_readonly(path)
 
     def unset_readonly(self, ref):
+        # type: (Text) -> None
+
         path = self.abspath(ref)
         if os.path.exists(path):
             self.unset_path_readonly(path)
 
     def clean_xattr_root(self):
+        # type: () -> None
+
         self.unlock_ref(u'/', unlock_parent=False)
         try:
             self.remove_root_id()
@@ -179,6 +192,8 @@ class LocalClient(BaseClient):
         self.clean_xattr_folder_recursive(u'/')
 
     def clean_xattr_folder_recursive(self, path):
+        # type: (Text) -> None
+
         for child in self.get_children_info(path):
             locker = self.unlock_ref(child.path, unlock_parent=False)
             if child.remote_ref is not None:
@@ -188,15 +203,23 @@ class LocalClient(BaseClient):
                 self.clean_xattr_folder_recursive(child.path)
 
     def remove_root_id(self):
+        # type: () -> None
+
         self.remove_remote_id('/', name='ndriveroot')
 
     def set_root_id(self, value):
+        # type: (Text) -> None
+
         self.set_remote_id('/', value, name="ndriveroot")
 
     def get_root_id(self):
+        # type: () -> Text
+
         return self.get_remote_id('/', name='ndriveroot')
 
     def _remove_remote_id_windows(self, path, name='ndrive'):
+        # type: (Text, Text) -> None
+
         path_alt = path + ':' + name
         try:
             os.remove(path_alt)
@@ -211,11 +234,12 @@ class LocalClient(BaseClient):
 
     @staticmethod
     def _remove_remote_id_unix(path, name='ndrive'):
+        # type: (Text, Text) -> None
+
+        if sys.platform == 'linux2':
+            name = 'user.' + name
         try:
-            if AbstractOSIntegration.is_mac():
-                xattr.removexattr(path, name)
-            else:
-                xattr.removexattr(path, 'user.' + name)
+            xattr.removexattr(path, name)
         except IOError as exc:
             # EPROTONOSUPPORT: protocol not supported (xattr)
             # ENODATA: no data available
@@ -223,11 +247,13 @@ class LocalClient(BaseClient):
                 raise exc
 
     def remove_remote_id(self, ref, name='ndrive'):
+        # type: (Text, Text) -> None
+
         path = self.abspath(ref)
         log.trace('Removing xattr %r from %r', name, path)
         locker = self.unlock_path(path, False)
         func = (self._remove_remote_id_windows
-                if AbstractOSIntegration.is_windows()
+                if sys.platform == 'win32'
                 else self._remove_remote_id_unix)
         try:
             func(path, name=name)
@@ -240,34 +266,45 @@ class LocalClient(BaseClient):
             self.lock_path(path, locker)
 
     def unset_folder_icon(self, ref):
-        """ Unset the red icon. """
+        # type: (Text) -> None
+        """ Unset the folder icon. """
 
-        desktop_ini_file_path = os.path.join(self.abspath(ref), "desktop.ini")
-        if AbstractOSIntegration.is_mac():
-            desktop_ini_file_path = os.path.join(self.abspath(ref), "Icon\r")
-        if os.path.exists(desktop_ini_file_path):
-            os.remove(desktop_ini_file_path)
+        if sys.platform == 'darwin':
+            meta_file = os.path.join(self.abspath(ref), 'Icon\r')
+        elif sys.platform == 'win32':
+            meta_file = os.path.join(self.abspath(ref), 'desktop.ini')
+        else:
+            return
+
+        try:
+            os.remove(meta_file)
+        except OSError:
+            pass
 
     def has_folder_icon(self, ref):
-        target_folder = self.abspath(ref)
-        if AbstractOSIntegration.is_mac():
-            meta_file = os.path.join(target_folder, "Icon\r")
-            return os.path.exists(meta_file)
-        if AbstractOSIntegration.is_windows():
-            meta_file = os.path.join(target_folder, "desktop.ini")
-            return os.path.exists(meta_file)
-        return False
+        # type: (Text) -> bool
+        """ Check if the folder icon is set. """
+
+        if sys.platform == 'darwin':
+            meta_file = os.path.join(self.abspath(ref), 'Icon\r')
+        elif sys.platform == 'win32':
+            meta_file = os.path.join(self.abspath(ref), 'desktop.ini')
+        else:
+            return False
+
+        return os.path.exists(meta_file)
 
     def set_folder_icon(self, ref, icon):
         if icon is None:
             return
-        if AbstractOSIntegration.is_windows():
-            self.set_folder_icon_win32(ref, icon)
-        elif AbstractOSIntegration.is_mac():
+
+        if sys.platform == 'darwin':
             self.set_folder_icon_darwin(ref, icon)
+        elif sys.platform == 'win32':
+            self.set_folder_icon_win32(ref, icon)
 
     def set_folder_icon_win32(self, ref, icon):
-        """ Configure red color icon for a folder Windows / macOS. """
+        """ Configure red color icon for a folder Windows. """
 
         # Desktop.ini file content for Windows 7+.
         ini_file_content = """
@@ -281,8 +318,8 @@ FolderType=Generic
         desktop_ini_content = ini_file_content.format(icon=icon)
 
         # Create the desktop.ini file inside the ReadOnly shared folder.
-        created_ini_file_path = os.path.join(self.abspath(ref), 'desktop.ini')
-        attrib_command_path = self.abspath(ref)
+        os_path = self.abspath(ref)
+        created_ini_file_path = os.path.join(os_path, 'desktop.ini')
         if not os.path.exists(created_ini_file_path):
             try:
                 with open(created_ini_file_path, 'w') as create_file:
@@ -298,20 +335,23 @@ FolderType=Generic
                                        win32con.FILE_ATTRIBUTE_SYSTEM)
             win32api.SetFileAttributes(created_ini_file_path,
                                        win32con.FILE_ATTRIBUTE_HIDDEN)
+
         # Windows folder use READ_ONLY flag as a customization flag ...
         # https://support.microsoft.com/en-us/kb/326549
-        win32api.SetFileAttributes(attrib_command_path,
-                                   win32con.FILE_ATTRIBUTE_READONLY)
+        win32api.SetFileAttributes(os_path, win32con.FILE_ATTRIBUTE_READONLY)
 
     @staticmethod
     def _read_data(file_path):
-        """ The data file contains the mac icons. """
+        # type: (Text) -> bytes
+        """ The data file contains macOS icons. """
 
         with open(file_path, 'rb') as dat:
             return dat.read()
 
     @staticmethod
     def _get_icon_xdata():
+        # type: () -> List[int]
+
         entry_size = 32
         icon_flag_index = 8
         icon_flag_value = 4
@@ -320,6 +360,7 @@ FolderType=Generic
         return result
 
     def set_folder_icon_darwin(self, ref, icon):
+        # type: (Text, Text) -> None
         """
         macOS: configure a folder with a given custom icon
             1. Read the com.apple.ResourceFork extended attribute from the icon file
@@ -328,6 +369,7 @@ FolderType=Generic
             4. Set extended attributes com.apple.FinderInfo & com.apple.ResourceFork for icon file (name: Icon\r)
             5. Hide the icon file (name: Icon\r)
         """
+
         try:
             target_folder = self.abspath(ref)
             # Generate the value for 'com.apple.FinderInfo'
@@ -349,22 +391,26 @@ FolderType=Generic
             log.exception('Impossible to set the folder icon')
 
     def set_remote_id(self, ref, remote_id, name='ndrive'):
+        # type: (Text, Text, Text) -> None
+
         if not isinstance(remote_id, bytes):
             remote_id = unicodedata.normalize('NFC', remote_id).encode('utf-8')
-        # Can be move to another class
+
         path = self.abspath(ref)
         log.trace('Setting xattr %s with value %r on %r', name, remote_id, path)
         locker = self.unlock_path(path, False)
-        if AbstractOSIntegration.is_windows():
+        if sys.platform == 'win32':
             path_alt = path + ':' + name
             try:
                 if not os.path.exists(path):
                     raise NotFound()
-                stat = os.stat(path)
+
+                stat_ = os.stat(path)
                 with open(path_alt, 'w') as f:
                     f.write(remote_id)
+
                 # Avoid time modified change
-                os.utime(path, (stat.st_atime, stat.st_mtime))
+                os.utime(path, (stat_.st_atime, stat_.st_mtime))
             except IOError as e:
                 # Should not happen
                 if e.errno == os.errno.EACCES:
@@ -376,35 +422,35 @@ FolderType=Generic
                     raise e
             finally:
                 self.lock_path(path, locker)
-        else:
-            try:
-                stat = os.stat(path)
-                if AbstractOSIntegration.is_mac():
-                    xattr.setxattr(path, name, remote_id)
-                else:
-                    xattr.setxattr(path, 'user.' + name, remote_id)
-                os.utime(path, (stat.st_atime, stat.st_mtime))
-            finally:
-                self.lock_path(path, locker)
+            return
+
+        if sys.platform == 'linux2':
+            name = 'user.' + name
+        try:
+            stat_ = os.stat(path)
+            xattr.setxattr(path, name, remote_id)
+            os.utime(path, (stat_.st_atime, stat_.st_mtime))
+        finally:
+            self.lock_path(path, locker)
 
     def get_remote_id(self, ref, name='ndrive'):
-        # type: (unicode, unicode) -> Union[unicode, None]
-        # Can be move to another class
-        path = self.abspath(ref)
-        return self.get_path_remote_id(path, name)
+        # type: (Text, Text) -> Union[Text, None]
+
+        return self.get_path_remote_id(self.abspath(ref), name)
 
     @staticmethod
     def get_path_remote_id(path, name='ndrive'):
-        # type: (unicode, unicode) -> Union[unicode, None]
-        if AbstractOSIntegration.is_windows():
+        # type: (Text, Text) -> Union[Text, None]
+
+        if sys.platform == 'win32':
             path += ':' + name
             try:
                 with open(path) as f:
                     return unicode(f.read(), 'utf-8')
-            except:
+            except (IOError, OSError):
                 return None
 
-        if AbstractOSIntegration.is_linux():
+        if sys.platform == 'linux2':
             name = 'user.' + name
         try:
             value = xattr.getxattr(path, name)
@@ -413,14 +459,18 @@ FolderType=Generic
             return None
 
     def get_info(self, ref, raise_if_missing=True):
+        # type: (Text, bool) -> Union[FileInfo, None]
+
         if isinstance(ref, bytes):
             ref = unicode(ref)
+
         os_path = self.abspath(ref)
         if not os.path.exists(os_path):
             if raise_if_missing:
                 err = 'Could not find file into {!r}: ref={!r}, os_path={!r}'
                 raise NotFound(err.format(self.base_folder, ref, os_path))
             return None
+
         folderish = os.path.isdir(os_path)
         stat_info = os.stat(os_path)
         size = 0 if folderish else stat_info.st_size
@@ -429,6 +479,7 @@ FolderType=Generic
         except ValueError, e:
             log.error(str(e) + "file path: %s. st_mtime value: %s" % (str(os_path), str(stat_info.st_mtime)))
             mtime = datetime.datetime.utcfromtimestamp(0)
+
         # TODO Do we need to load it everytime ?
         remote_ref = self.get_remote_id(ref)
         # On unix we could use the inode for file move detection but that won't
@@ -447,6 +498,7 @@ FolderType=Generic
         local_path,
         remote_digest_algorithm=None,
     ):
+        # type: (Text, Text, Text, Optional[Text]) -> bool
         """
         Compare 2 document's digests.
 
@@ -460,6 +512,7 @@ FolderType=Generic
 
         if local_digest == remote_digest:
             return True
+
         if remote_digest_algorithm is None:
             remote_digest_algorithm = guess_digest_algorithm(remote_digest)
         if remote_digest_algorithm == self._digest_func:
@@ -470,11 +523,13 @@ FolderType=Generic
         return digest == remote_digest
 
     def get_content(self, ref):
+        # type: (Text) -> bytes
+
         with open(self.abspath(ref), 'rb') as f:
             return f.read()
 
     def is_ignored(self, parent_ref, file_name):
-        # type: (unicode, unicode) -> bool
+        # type: (Text, Text) -> bool
         """ Note: added parent_ref to be able to filter on size if needed. """
 
         file_name = file_name.lower()
@@ -483,7 +538,7 @@ FolderType=Generic
                 or file_name.startswith(Options.ignored_prefixes)):
             return True
 
-        if AbstractOSIntegration.is_windows():
+        if sys.platform == 'win32':
             # NXDRIVE-465: ignore hidden files on Windows
             ref = self.get_children_ref(parent_ref, file_name)
             path = self.abspath(ref)
@@ -509,11 +564,15 @@ FolderType=Generic
 
     @staticmethod
     def get_children_ref(parent_ref, name):
+        # type: (Text, Text) -> Text
+
         if parent_ref == u'/':
             return parent_ref + name
         return parent_ref + u'/' + name
 
     def get_children_info(self, ref):
+        # type: (Text) -> List[FileInfo]
+
         os_path = self.abspath(ref)
         result = []
         children = os.listdir(os_path)
@@ -537,24 +596,33 @@ FolderType=Generic
 
     @staticmethod
     def get_parent_ref(ref):
+        # type: (Text) -> Union[Text, None]
+
         if ref == '/':
             return None
+
         parent = ref.rsplit(u'/', 1)[0]
         if parent is None:
             parent = '/'
         return parent
 
-    def unlock_ref(self, ref, unlock_parent=True):
-        path = self.abspath(ref)
+    def unlock_ref(self, ref, unlock_parent=True, is_abs=False):
+        # type: (Text, bool, bool) -> int
+
+        path = ref if is_abs else self.abspath(ref)
         return self.unlock_path(path, unlock_parent)
 
-    def lock_ref(self, ref, locker):
-        path = self.abspath(ref)
+    def lock_ref(self, ref, locker, is_abs=False):
+        # type: (Text, int, bool) -> int
+
+        path = ref if is_abs else self.abspath(ref)
         return self.lock_path(path, locker)
 
     def make_folder(self, parent, name):
-        locker = self.unlock_ref(parent, False)
+        # type: (Text, Text) -> Text
+
         os_path, name = self._abspath_deduped(parent, name)
+        locker = self.unlock_ref(parent, unlock_parent=False)
         try:
             os.mkdir(os_path)
             # Name should be the actual name of the folder created locally
@@ -567,11 +635,8 @@ FolderType=Generic
 
     @staticmethod
     def make_tree(path):
-        """
-        Recursive directory creation.
-
-        :param str path: The absolute path to create.
-        """
+        # type: (Text) -> None
+        """ Recursive directory creation. """
 
         try:
             os.makedirs(path)
@@ -581,10 +646,12 @@ FolderType=Generic
                 raise exc
 
     def duplicate_file(self, ref):
+        # type: (Text) -> Text
+
         parent = os.path.dirname(ref)
         name = os.path.basename(ref)
-        locker = self.unlock_ref(parent, False)
         os_path, name = self._abspath_deduped(parent, name)
+        locker = self.unlock_ref(os_path, unlock_parent=False, is_abs=True)
         if parent == u"/":
             duplicated_file = u"/" + name
         else:
@@ -596,13 +663,15 @@ FolderType=Generic
             e.duplicated_file = duplicated_file
             raise e
         finally:
-            self.lock_ref(parent, locker)
+            self.lock_ref(os_path, locker, is_abs=True)
 
     def make_file(self, parent, name, content=None):
-        locker = self.unlock_ref(parent, False)
+        # type: (Text, Text, Optional[bytes]) -> Text
+
         os_path, name = self._abspath_deduped(parent, name)
+        locker = self.unlock_ref(parent, unlock_parent=False)
         try:
-            with open(os_path, "wb") as f:
+            with open(os_path, 'wb') as f:
                 if content:
                     f.write(content)
             if parent == u"/":
@@ -612,6 +681,8 @@ FolderType=Generic
             self.lock_ref(parent, locker)
 
     def get_new_file(self, parent, name):
+        # type: (Text, Text) -> Tuple[Text, Text, Text]
+
         os_path, name = self._abspath_deduped(parent, name)
         if parent == u"/":
             path = u"/" + name
@@ -620,6 +691,8 @@ FolderType=Generic
         return path, os_path, name
 
     def update_content(self, ref, content, xattr_names=('ndrive',)):
+        # type: (Text, bytes, Tuple[Text]) -> None
+
         xattrs = {name: self.get_remote_id(ref, name=name)
                   for name in xattr_names}
 
@@ -631,13 +704,15 @@ FolderType=Generic
                 self.set_remote_id(ref, value, name=name)
 
     def delete(self, ref):
-        locker = self.unlock_ref(ref)
+        # type: (Text) -> None
+
         os_path = self.abspath(ref)
-        if not self.exists(ref):
+        if not os.path.exists(os_path):
             return
 
-        # Remove the \\?\ prefix, specific to Windows
-        os_path = os_path.lstrip('\\\\?\\')
+        if sys.platform == 'win32':
+            # Remove the '\\?\' prefix
+            os_path = os_path.lstrip('\\\\?\\')
 
         log.trace('Trashing %r', os_path)
 
@@ -645,6 +720,7 @@ FolderType=Generic
         if not isinstance(os_path, bytes):
             os_path = os_path.encode(sys.getfilesystemencoding() or 'utf-8')
 
+        locker = self.unlock_ref(os_path, is_abs=True)
         try:
             send2trash(os_path)
         except OSError:
@@ -652,15 +728,17 @@ FolderType=Generic
             self.delete_final(ref)
         finally:
             # Don't want to unlock the current deleted
-            self.lock_ref(ref, locker & 2)
+            self.lock_ref(os_path, locker & 2, is_abs=True)
 
     def delete_final(self, ref):
+        # type: (Text) -> None
+
         locker = 0
         parent_ref = None
         try:
-            if ref is not '/':
+            if ref != '/':
                 parent_ref = os.path.dirname(ref)
-                locker = self.unlock_ref(parent_ref, False)
+                locker = self.unlock_ref(parent_ref, unlock_parent=False)
             self.unset_readonly(ref)
             os_path = self.abspath(ref)
             if os.path.isfile(os_path):
@@ -672,25 +750,25 @@ FolderType=Generic
                 self.lock_ref(parent_ref, locker)
 
     def exists(self, ref):
-        os_path = self.abspath(ref)
-        return os.path.exists(os_path)
+        # type: (Text) -> bool
+
+        return os.path.exists(self.abspath(ref))
 
     def check_writable(self, ref):
-        os_path = self.abspath(ref)
-        return os.access(os_path, os.W_OK)
+        # type: (Text) -> bool
+
+        return os.access(self.abspath(ref), os.W_OK)
 
     def rename(self, ref, to_name):
-        """Rename a local file or folder
+        # type: (Text, Text) -> FileInfo
+        """ Rename a local file or folder. """
 
-        Return the actualized info object.
-
-        """
         new_name = safe_filename(to_name)
         source_os_path = self.abspath(ref)
         parent = ref.rsplit(u'/', 1)[0]
         old_name = ref.rsplit(u'/', 1)[1]
         parent = u'/' if parent == '' else parent
-        locker = self.unlock_ref(ref)
+        locker = self.unlock_ref(source_os_path, is_abs=True)
         try:
             # Check if only case renaming
             if (old_name != new_name
@@ -701,7 +779,7 @@ FolderType=Generic
                     temp_path = os.tempnam(
                         self.abspath(parent),
                         self.CASE_RENAME_PREFIX + old_name + '_')
-                if AbstractOSIntegration.is_windows():
+                if sys.platform == 'win32':
                     ctypes.windll.kernel32.SetFileAttributesW(
                         unicode(temp_path), 2)
                 os.rename(source_os_path, temp_path)
@@ -713,40 +791,43 @@ FolderType=Generic
                     parent, new_name, old_name)
             if old_name != new_name:
                 os.rename(source_os_path, target_os_path)
-            if AbstractOSIntegration.is_windows():
+            if sys.platform == 'win32':
                 # See http://msdn.microsoft.com/en-us/library/aa365535%28v=vs.85%29.aspx
                 ctypes.windll.kernel32.SetFileAttributesW(
                     unicode(target_os_path), 128)
             new_ref = self.get_children_ref(parent, new_name)
             return self.get_info(new_ref)
         finally:
-            self.lock_ref(ref, locker & 2)
+            self.lock_ref(source_os_path, locker & 2, is_abs=True)
 
     def move(self, ref, new_parent_ref, name=None):
-        """Move a local file or folder into another folder
+        # type: (Text, Text, Optional[Text]) -> FileInfo
+        """ Move a local file or folder into another folder. """
 
-        Return the actualized info object.
-
-        """
         if ref == u'/':
-            raise ValueError("Cannot move the toplevel folder.")
-        locker = self.unlock_ref(ref)
-        new_locker = self.unlock_ref(new_parent_ref, False)
-        source_os_path = self.abspath(ref)
+            raise ValueError('Cannot move the toplevel folder.')
+
         name = name if name is not None else ref.rsplit(u'/', 1)[1]
+        filename = self.abspath(ref)
         target_os_path, new_name = self._abspath_deduped(new_parent_ref, name)
+        locker = self.unlock_ref(filename, is_abs=True)
+        parent = os.path.dirname(target_os_path)
+        new_locker = self.unlock_ref(parent, unlock_parent=False, is_abs=True)
         try:
-            os.rename(source_os_path, target_os_path)
+            os.rename(filename, target_os_path)
             new_ref = self.get_children_ref(new_parent_ref, new_name)
             return self.get_info(new_ref)
         finally:
-            self.lock_ref(ref, locker & 2)
-            self.lock_ref(new_parent_ref, locker & 1 | new_locker)
+            self.lock_ref(filename, locker & 2, is_abs=True)
+            self.lock_ref(parent, locker & 1 | new_locker, is_abs=True)
 
     def is_inside(self, abspath):
+        # type: (Text) -> bool
+
         return abspath.startswith(self.base_folder)
 
     def get_path(self, abspath):
+        # type: (Text) -> Text
         """ Relative path to the local client from an absolute OS path. """
 
         if isinstance(abspath, bytes):
@@ -758,30 +839,36 @@ FolderType=Generic
         return path.replace(os.path.sep, '/')
 
     def abspath(self, ref):
-        """Absolute path on the operating system"""
+        # type: (Text) -> Text
+        """ Absolute path on the operating system. """
+
         if not ref.startswith(u'/'):
             raise ValueError(
                 'LocalClient expects ref starting with "/"', locals())
+
         path_suffix = ref[1:].replace('/', os.path.sep)
         path = normalized_path(os.path.join(self.base_folder, path_suffix))
         return safe_long_path(path)
 
     def _abspath_safe(self, parent, orig_name):
-        """Absolute path on the operating system with deduplicated names"""
-        # make name safe by removing invalid chars
+        # type: (Text, Text) -> Text
+        """ Absolute path on the operating system with deduplicated names. """
+
+        # Make name safe by removing invalid chars
         name = safe_filename(orig_name)
 
-        # decompose the name into actionable components
+        # Decompose the name into actionable components
         name, suffix = os.path.splitext(name)
-        os_path = self.abspath(os.path.join(parent, name + suffix))
-        return os_path
+        return self.abspath(os.path.join(parent, name + suffix))
 
     def _abspath_deduped(self, parent, orig_name, old_name=None):
-        """Absolute path on the operating system with deduplicated names"""
-        # make name safe by removing invalid chars
+        # type: (Text, Text, Optional[Text]) -> Tuple[Text, Text]
+        """ Absolute path on the operating system with deduplicated names. """
+
+        # Make name safe by removing invalid chars
         name = safe_filename(orig_name)
 
-        # decompose the name into actionable components
+        # Decompose the name into actionable components
         name, suffix = os.path.splitext(name)
 
         for _ in range(1000):