+++ tools/integration_tests_setup.py
@@ -179,6 +179,9 @@ def set_environment(server_url, engine):
     os.environ['NXDRIVE_TEST_USER'] = "Administrator"
     os.environ['NXDRIVE_TEST_PASSWORD'] = "Administrator"
 
+    # Convenient way to try a specific test without having to abort and start a new job
+    os.environ['SPECIFIC_TEST'] = ''
+
 
 def clean_pyc(dir_):
     for root, dirnames, filenames in os.walk(dir_):